created separate .py file for classes
