+++ Game/run_game.py
@@ -1,0 +1,119 @@
+import pygame
+from characters import Player, Projectile, Enemy
+pygame.init()
+
+win = pygame.display.set_mode((500, 480))
+bg = pygame.image.load('bg.jpg')
+pygame.display.set_caption("First Game")
+
+clock = pygame.time.Clock()
+
+bulletSound = pygame.mixer.Sound('bullet.mp3')
+hitSound = pygame.mixer.Sound('hit.mp3')
+
+pygame.mixer.music.load('music.mp3')
+pygame.mixer.music.play(-1)
+
+score = 0
+
+
+def redraw_game_window():
+    win.blit(bg, (0, 0))
+    text = font.render('Score: ' + str(score), 1, (0, 0, 0))
+    win.blit(text, (350, 10))
+    man.draw(win)
+    goblin.draw(win)
+    for bullet in bullets:
+        bullet.draw(win)
+    
+    pygame.display.update()
+
+
+# mainloop
+font = pygame.font.SysFont('comicsans', 30, True)
+man = Player(200, 410, 64,64)
+goblin = Enemy(100, 410, 64, 64, 450)
+shootLoop = 0
+bullets = []
+run = True
+while run:
+    clock.tick(27)
+
+    if goblin.visible:
+        if man.hitbox[1] < goblin.hitbox[1] + goblin.hitbox[3] and man.hitbox[1] + man.hitbox[3] > goblin.hitbox[1]:
+            if man.hitbox[0] + man.hitbox[2] > goblin.hitbox[0] and man.hitbox[0] < goblin.hitbox[0] + goblin.hitbox[2]:
+                man.hit(win)
+                score -= 5
+
+    if shootLoop > 0:
+        shootLoop += 1
+    if shootLoop > 3:
+        shootLoop = 0
+    
+    for event in pygame.event.get():
+        if event.type == pygame.QUIT:
+            run = False
+        
+    for bullet in bullets:
+        if bullet.y - bullet.radius < goblin.hitbox[1] + goblin.hitbox[3] and bullet.y + bullet.radius > goblin.hitbox[1]:
+            if bullet.x + bullet.radius > goblin.hitbox[0] and bullet.x - bullet.radius < goblin.hitbox[0] + goblin.hitbox[2]:
+                hitSound.play()
+                goblin.hit()
+                score += 1
+                bullets.pop(bullets.index(bullet))
+                
+        if 0 < bullet.x < 500:
+            bullet.x += bullet.vel
+        else:
+            bullets.pop(bullets.index(bullet))
+
+    keys = pygame.key.get_pressed()
+
+    if keys[pygame.K_SPACE] and shootLoop == 0:
+        bulletSound.play()
+        if man.left:
+            facing = -1
+        else:
+            facing = 1
+            
+        if len(bullets) < 5:
+            bullets.append(Projectile(round(man.x + man.width // 2), round(man.y + man.height//2), 6, (0, 0, 0), facing))
+
+        shootLoop = 1
+
+    if keys[pygame.K_LEFT] and man.x > man.vel:
+        man.x -= man.vel
+        man.left = True
+        man.right = False
+        man.standing = False
+    elif keys[pygame.K_RIGHT] and man.x < 500 - man.width - man.vel:
+        man.x += man.vel
+        man.right = True
+        man.left = False
+        man.standing = False
+    else:
+        man.standing = True
+        man.walkCount = 0
+        
+    if not man.isJump:
+        if keys[pygame.K_UP]:
+            man.isJump = True
+            man.right = False
+            man.left = False
+            man.walkCount = 0
+    else:
+        if man.jumpCount >= -10:
+            neg = 1
+            if man.jumpCount < 0:
+                neg = -1
+            man.y -= (man.jumpCount ** 2) * 0.5 * neg
+            man.jumpCount -= 1
+        else:
+            man.isJump = False
+            man.jumpCount = 10
+            
+    redraw_game_window()
+
+pygame.quit()
+
+
